Added solution - LeetHub

+++ Square root of a number - GFG/square-root-of-a-number.cpp
@@ -12,7 +12,7 @@ class Solution{
     long long int floorSqrt(long long int x) 
     {
         // Your code goes here
-        long long sqroot;
+        int sqroot;
    for(int i=0;i*i<=x;i++){
    sqroot=i;
    }
